feat: project showcase and app router
