+++ noahgiboney.app/src/app/home.css
@@ -206,3 +206,4 @@
   background-color: #655699;
   color: white;
 }
+
