+++ noahgiboney.app/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
 import { Josefin_Sans } from 'next/font/google'
 import './globals.css'
-import Navbar from '@/components/navbar/navbar';
+import Navbar from '@/components/Navbar/navbar';
 import Footer from '@/components/Footer/footer';
 
 const inter = Josefin_Sans({ subsets: ['latin'] })
