+++ noahgiboney.app/src/app/virtualworld/page.tsx
@@ -1,0 +1,13 @@
+// import linkedinIcon from './images/icons/linkedin.png'; // Adjust the path as needed
+// import githubIcon from './images/icons/github.png';    // Adjust the path as needed
+// import mailIcon from './images/icons/mail.png';       // Adjust the path as needed
+import PortfolioSection from '@/components/Portfolio/portfolio';
+
+
+export default function Home() {
+  return (
+    <div>
+        <p>Hello this page will be for my virtual world project</p>
+    </div>
+  );
+}
